full game functionality including decreasing ingredients with sales

+++ Player.cs
@@ -9,17 +9,17 @@ namespace LemonadeStand
     class Player
     {
         //member variables
-        public int cashBalance;
+        public double cashBalance;
 
-        public int cupsSugar;
-        public int manyLemons;
-        public int poundsOfIce;
-        public int lemonadeRecipe;
+        public double cupsSugar;
+        public double manyLemons;
+        public double poundsOfIce;
+        public double lemonadeRecipe;
         public int priceOfLemonade;
 
-        public int pitcherSugar;
-        public int pitcherLemons;
-        public int pitcherIce;
+        public double pitcherSugar;
+        public double pitcherLemons;
+        public double pitcherIce;
 
         public double cupsToSell;
 
@@ -33,7 +33,7 @@ namespace LemonadeStand
         }
         //member methods
 
-        public int CreateRecipe()
+        public double CreateRecipe()
         {
             Console.WriteLine("How many cups of sugar would you like per pitcher?");
             cupsSugar = Convert.ToInt32(Console.ReadLine());
@@ -52,7 +52,7 @@ namespace LemonadeStand
             return priceOfLemonade;
         }
 
-        public double PredictProfit(int itemOne)
+        public double PredictProfit(double itemOne)
         {
             if (cupsToSell <= itemOne)
             {
@@ -68,7 +68,7 @@ namespace LemonadeStand
                 return potentialProfit;
             }
         }
-        public double MakeLemonade(int itemOne, int itemTwo, int itemThree)
+        public double MakeLemonade(double itemOne, double itemTwo, double itemThree)
         {
             //calculates how much lemonade there will be for the day based on 
             //quantity of ingredients in inventory and recipe
